i3blocks: weather: Specify city by name.
It used to be the case that openweathermap would ignore any supplied country code
and pick the most common city with that name.
This is not true any more, it uses the country code so we can just specify the city
that way.

It is ncessary to quote the macro name in i3blocks config because the expansion
contains a comma, which is treated as an argument separator if not quoted.